refac: use vite to replace webpack for dev and build
