+++ .commitlintrc.ts
@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   parserPreset: {
     parserOpts: {
       headerPattern: /^(\w*)(?:\((.*)\))?:[ ]?(.*)$/,
