+++ .eslintignore
@@ -1,7 +1,6 @@
 docs
 lib/assets/libs/sequence
-example/
-scripts/
+src/
 commitlint.config.js
 jest.config.js
 lint-staged.config.js
