--- .eslintrc.cjs
@@ -1,7 +1,0 @@
-/* eslint-env node */
-module.exports = {
-  extends: ["eslint:recommended", "plugin:@typescript-eslint/recommended"],
-  parser: "@typescript-eslint/parser",
-  plugins: ["@typescript-eslint"],
-  root: true,
-};
