+++ .lintstagedrc
@@ -1,0 +1,4 @@
+{
+  "*.ts": ["eslint --fix"],
+  "*.{html,css}": ["stylelint \"/**/*.css\" --fix"]
+}
