--- example/index.html
@@ -1,54 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Muya Development</title>
-  <style>
-    @supports(-moz-appearance:none) {
-      /* CSS code */
-      html {
-        position: fixed;
-      }
-    }
-    html,
-    body {
-      margin: 0;
-      padding: 0;
-    }
-    #editor {
-      box-sizing: border-box;
-      width: 100%;
-      height: 100vh;
-      overflow-y: auto;
-
-      font-size: 16px;
-      
-      outline: none;
-    }
-    #search {
-      width: 200px;
-      height: 18px;
-
-      border: 1px slid #eee;
-      border-radius: 4px;
-      outline: none;
-    }
-  </style>
-</head>
-<body>
-  <div class="tools">
-    <button id="undo">UNDO</button>
-    <button id="redo">REDO</button>
-    <input type="text" id="search">
-    <button id="previous">Previous</button>
-    <button id="next">Next</button>
-    <input type="text" id="replace">
-    <button id="single">Single</button>
-    <button id="all">All</button>
-    <button id="set-content">Set Content</button>
-    <button id="select-all">Select All</button>
-  </div>
-  <div id="editor"></div>
-</body>
-</html>
