+++ index.html
@@ -1,0 +1,25 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8" />
+  <link rel="icon" type="image/svg+xml" href="/logo.jpg" />
+  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+  <title>Muya Development</title>
+</head>
+<body>
+  <div class="tools">
+    <button id="undo">UNDO</button>
+    <button id="redo">REDO</button>
+    <input type="text" id="search">
+    <button id="previous">Previous</button>
+    <button id="next">Next</button>
+    <input type="text" id="replace">
+    <button id="single">Single</button>
+    <button id="all">All</button>
+    <button id="set-content">Set Content</button>
+    <button id="select-all">Select All</button>
+  </div>
+  <div id="editor"></div>
+  <script type="module" src="/src/main.ts"></script>
+</body>
+</html>
