+++ jest.config.json
@@ -1,0 +1,7 @@
+{
+  "clearMocks": true,
+  "coverageDirectory": "coverage",
+  "testEnvironment": "node",
+  "testMatch": ["**/__tests__/**/*.js?(x)", "**/?(*.)+(spec|test).js?(x)"],
+  "testPathIgnorePatterns": ["/node_modules/"]
+}
