+++ lib/assets/styles/exportStyle.css
@@ -4,6 +4,7 @@
 
 .footnotes {
   font-size: .85em;
+
   opacity: .8;
 }
 
@@ -13,15 +14,16 @@
 
 .footnotes .footnote-back {
   position: absolute;
-  font-family: initial;
   top: .2em;
   right: 1em;
+
+  font-family: initial;
   text-decoration: none;
 }
 
 .inline-math.invalid,
 .multiple-math.invalid {
-  color: rgb(255, 105, 105);
+  color: rgb(255 105 105);
 }
 
 .toc-container {
@@ -29,9 +31,10 @@
 }
 
 .toc-container .toc-title {
+  margin-bottom: 0;
+
   font-weight: 700;
   font-size: 1.2em;
-  margin-bottom: 0;
 }
 
 .toc-container li,
@@ -52,6 +55,7 @@
   color: inherit;
   text-decoration: none;
 }
+
 .toc-container ul li span a:hover {
   color: inherit;
   text-decoration: none;
@@ -61,5 +65,6 @@
   flex: 1;
   height: 0.65em;
   margin: 0 10px;
+
   border-bottom: 2px dotted black;
 }
