+++ lib/assets/styles/headerFooterStyle.css
@@ -1,5 +1,5 @@
 :root {
-  --footerHeaderBorderColor: #1c1c1c;
+  --footer-header-border-color: #1c1c1c;
 }
 
 table.page-container,
@@ -16,6 +16,7 @@ table.page-container > tfoot > tr,
 table.page-container > tfoot > tr > th,
 table.page-container > tfoot > tr > td  {
   position: relative !important;
+
   margin: 0 !important;
   padding: 0 !important;
 }
@@ -25,36 +26,53 @@ table.page-container > tfoot > tr > td  {
 .page-footer .hf-container {
   display: flex;
   justify-content: space-between;
-  font-size: 0.75em;
+
   font-weight: 400;
+  font-size: 0.75em;
+}
+
+.page-footer-fake .hf-container {
+  margin-top: 16px;
+
+  visibility: hidden;
 }
 
 .page-header {
   display: table-header-group;
 }
+
 .page-header .hf-container {
   margin-bottom: 16px;
 }
+
 .page-header.styled .hf-container {
   padding-bottom: 1px;
-  border-bottom: 1px solid var(--footerHeaderBorderColor);
+
+  border-bottom: 1px solid var(--footer-header-border-color);
 }
+
 .page-header .hf-container > div {
   flex: 1;
   max-height: 100px;
   overflow: hidden;
 }
+
 .page-header .header-content-left {
-  text-align: left;
   margin-right: 4px;
+
+  text-align: left;
 }
+
 .page-header .header-content {
   text-align: center;
 }
+
 .page-header .header-content-right {
-  text-align: right;
   margin-left: 4px;
+
+  text-align: right;
 }
+
 .page-header.single .header-content-left,
 .page-header.single .header-content-right {
   display: none;
@@ -63,36 +81,43 @@ table.page-container > tfoot > tr > td  {
 .page-footer-fake {
   display: table-footer-group;
 }
-.page-footer-fake .hf-container {
-  margin-top: 16px;
-  visibility: hidden;
-}
+
 .page-footer {
   position: fixed;
+  right: 0;
   bottom: 0;
   left: 0;
-  right: 0;
 }
+
 .page-footer.styled .hf-container {
   padding-top: 1px;
-  border-top: 1px solid var(--footerHeaderBorderColor);
+
+  border-top: 1px solid var(--footer-header-border-color);
 }
+
 .page-footer .hf-container > div {
   flex: 1;
-  white-space: nowrap;
   overflow: hidden;
+
+  white-space: nowrap;
 }
+
 .page-footer .footer-content-left {
-  text-align: left;
   margin-right: 14px;
+
+  text-align: left;
 }
+
 .page-footer .footer-content {
   text-align: center;
 }
+
 .page-footer .footer-content-right {
-  text-align: right;
   margin-left: 14px;
+
+  text-align: right;
 }
+
 .page-footer.single .footer-content-left,
 .page-footer.single .footer-content-right {
   display: none;
