+++ lib/assets/styles/index.css
@@ -125,11 +125,9 @@
 @keyframes loading {
   0% {
     transform: rotate(0deg);
-    transform: rotate(0deg);
   }
 
   100% {
-    transform: rotate(360deg);
     transform: rotate(360deg);
   }
 }
