+++ lib/assets/styles/prismjs/light.theme.css
@@ -21,13 +21,7 @@ pre.mu-code-block {
   word-wrap: normal;
   word-break: normal;
   word-spacing: normal;
-  -moz-tab-size: 4;
-  -o-tab-size: 4;
   tab-size: 4;
-
-  -webkit-hyphens: none;
-  -moz-hyphens: none;
-  -ms-hyphens: none;
   hyphens: none;
 }
 
@@ -104,6 +98,7 @@ pre.mu-code-block {
 .token.bold {
   font-weight: bold;
 }
+
 .token.italic {
   font-style: italic;
 }
