+++ lib/block/commonMark/atxHeading/index.ts
@@ -1,21 +1,21 @@
-// @ts-nocheck
 import Parent from "@muya/block/base/parent";
 import ScrollPage from "@muya/block/scrollPage";
 import { mixins } from "@muya/utils";
 import leafQueryBlock from "@muya/block/mixins/leafQueryBlock";
 import { IAtxHeadingState } from "../../../../types/state";
-import AtxHeadingContent from "@muya/block/content/atxHeadingContent";
-import LinkedList from "@muya/block/base/linkedList/linkedList";
+import Muya from "@muya/index";
+import Content from "@muya/block/base/content";
 
 interface IAtxHeadingMeta {
   level: number;
 }
 
 class AtxHeading extends Parent {
-  public children: LinkedList<AtxHeadingContent>;
+  public meta: IAtxHeadingMeta;
+
   static blockName = "atx-heading";
 
-  static create(muya, state) {
+  static create(muya: Muya, state: IAtxHeadingState) {
     const heading = new AtxHeading(muya, state);
 
     heading.append(
@@ -25,16 +25,14 @@ class AtxHeading extends Parent {
     return heading;
   }
 
-  public meta: IAtxHeadingMeta;
-
   get path() {
-    const { path: pPath } = this.parent;
-    const offset = this.parent.offset(this);
+    const { path: pPath } = this.parent!;
+    const offset = this.parent!.offset(this);
 
     return [...pPath, offset];
   }
 
-  constructor(muya, { meta }) {
+  constructor(muya: Muya, { meta }: IAtxHeadingState) {
     super(muya);
     this.tagName = `h${meta.level}`;
     this.meta = meta;
@@ -46,7 +44,7 @@ class AtxHeading extends Parent {
     return {
       name: "atx-heading",
       meta: this.meta,
-      text: this.children.head.text,
+      text: (this.children.head as Content).text,
     };
   }
 }
