+++ lib/utils/dompurify.ts
@@ -1,5 +1,7 @@
-import { sanitize, isValidAttribute } from "dompurify";
+import DOMPurify from "dompurify";
 import type { Config } from "dompurify";
+
+const { sanitize, isValidAttribute } = DOMPurify();
 
 export { isValidAttribute, Config };
