+++ lib/utils/snabbdom.ts
@@ -9,6 +9,7 @@ import {
   h as sh,
   toVNode as sToVNode,
 } from "snabbdom";
+import sToHTML from "snabbdom-to-html";
 
 export const patch = init([
   classModule,
@@ -22,7 +23,7 @@ export const patch = init([
 export const h = sh;
 export const toVnode = sToVNode;
 
-export const toHTML = require("snabbdom-to-html"); // helper function for convert vnode to HTML string
+export const toHTML = sToHTML; // helper function for convert vnode to HTML string
 export const htmlToVNode = (html: string) => {
   // helper function for convert html to vnode
   const wrapper = document.createElement("div");
