--- lint-staged.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  '*.ts': ['eslint --fix'],
-  '*.{html,css,sass,scss}': ['stylelint --allow-empty-input --fix']
-}
