+++ package.json
@@ -1,16 +1,16 @@
 {
   "name": "@marktext/muya",
   "version": "0.1.13",
-  "description": "Future markdown editor for web and desktop application development",
+  "description": "Future markdown editor",
   "main": "dist/muya.js",
+  "type": "module",
   "scripts": {
-    "dev": "webpack serve --config ./example/webpack.config.js --open",
-    "build": "npm run lint:tsc && webpack --config scripts/webpack.config.js",
+    "dev": "vite",
+    "build": "tsc && vite build",
     "test": "jest --env=jsdom",
-    "lint": "npm run lint:eslint && npm run lint:tsc",
-    "lint:eslint": "eslint lib",
-    "lint:tsc": "tsc --noEmit --skipLibCheck",
-    "lint:fix": "eslint --fix lib",
+    "lint:ts": "eslint lib",
+    "lint:ts:fix": "eslint --fix lib",
+    "lint:css": "stylelint \"**/*.css\"",
     "postinstall": "husky install",
     "prepublishOnly": "pinst --disable",
     "postpublish": "pinst --enable",
@@ -62,47 +62,22 @@
     "webfontloader": "^1.6.28"
   },
   "devDependencies": {
-    "@babel/core": "7.22.19",
-    "@babel/helper-string-parser": "^7.22.5",
-    "@babel/preset-env": "7.22.15",
-    "@babel/preset-typescript": "^7.22.15",
-    "@types/d3": "^7.4.0",
     "@types/dompurify": "^3.0.2",
     "@types/prismjs": "^1.26.0",
-    "@types/vfile-message": "^2.0.0",
     "@typescript-eslint/eslint-plugin": "^6.7.0",
     "@typescript-eslint/parser": "^6.7.0",
-    "babel-jest": "^29.7.0",
-    "babel-loader": "^9.1.3",
-    "browserify-zlib": "^0.2.0",
     "commitlint": "^17.7.1",
-    "cross-env": "^7.0.3",
-    "css-loader": "^6.8.1",
-    "dotenv-webpack": "^8.0.1",
     "eslint": "^8.49.0",
-    "file-loader": "^6.2.0",
-    "html-webpack-plugin": "^5.5.3",
     "husky": "8.0.3",
-    "imports-loader": "^4.0.1",
     "jest": "^29.7.0",
     "lint-staged": "^14.0.1",
-    "mini-css-extract-plugin": "^2.7.6",
-    "node-polyfill-webpack-plugin": "^2.0.1",
     "pinst": "^3.0.0",
-    "postcss": "^8.4.29",
-    "postcss-loader": "^7.3.3",
-    "postcss-preset-env": "^9.1.3",
-    "raw-loader": "^4.0.2",
-    "source-map-loader": "^4.0.1",
-    "style-loader": "^3.3.3",
     "stylelint": "^15.10.3",
     "stylelint-config-prettier": "^9.0.5",
     "stylelint-config-rational-order": "^0.1.2",
     "stylelint-config-standard": "^34.0.0",
     "stylelint-order": "^6.0.3",
     "typescript": "^5.2.2",
-    "webpack": "^5.88.2",
-    "webpack-cli": "^5.1.4",
-    "webpack-dev-server": "^4.15.1"
+    "vite": "^4.4.5"
   }
 }
