+++ tsconfig.json
@@ -1,21 +1,28 @@
 {
   "compilerOptions": {
     "allowSyntheticDefaultImports": true,
-    "target": "es2020",
-    "sourceMap": true,
-    "declaration": true,
-    "module": "ES2020",
-    "moduleResolution": "Node",
-    "lib": ["DOM", "ES2015"],
+    "target": "ES2020",
+    "useDefineForClassFields": true,
+    "module": "ESNext",
+    "lib": ["ES2020", "DOM", "DOM.Iterable"],
+    "skipLibCheck": true,
+
+    /* Bundler mode */
+    "moduleResolution": "bundler",
+    "allowImportingTsExtensions": true,
     "resolveJsonModule": true,
-    "baseUrl": ".",
+    "isolatedModules": true,
     "noEmit": true,
-    "noFallthroughCasesInSwitch": true,
+
+    /* Linting */
     "strict": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "noFallthroughCasesInSwitch": true,
     "paths": {
       "@muya/*": ["./lib/*"],
       "@types/*": ["./types/*"]
     }
   },
-  "include": ["./lib/**/*", "./types/**/*"],
+  "include": ["src", "types", "lib"],
 }
